fix: move back some V2 specific code
